Changed logic for choosing outgoing connection port
The selection of outgoing port depends on the nature of the deployment
of the libhtp proxy.  Different deployment scenarios include:
- explicitly set as the browser proxy
- operating as a transparent proxy (eg using linux TRPOXY)
- using a firewall redirect but with dport altered
- using a firewall redirect but with dport unaltered

The special case is case 4, where the connection is redirected to the
proxy at the firewall level, meaning the proxy receives the connection
as the remote server would.  In this case, the outgoing connection port
should be the same as the local connection properties.

In the other cases, the outgoing connection properties are the
remote_address and remote_port specified when creating the connection
parser, or the properties parsed from the HTTP headers.

Use a new flag: use_local_port to indicate which port to use.  When the
proxy deployment is case 4, the connection parser should be created with
this value set to true (htp_connp_open).

In future, this can be an enum type rather than just a boolean flag,
to indicate the deployment scenario, to cause small changes in the
behaviour of libhtp, in much the same style as server personalities
change some behaviours in libhtp.