In htp_log, set the htp_log_t::tx field
The htp_log_t structure has a pointer to the transaction in by which
the message was raised. However, the field is never set.

This patch sets it based on either the htp_connp_t::in_tx or the
htp_connp_t::out_tx pointers. The former has preference over the
latter.