Improve request line parsing with leading spaces
Leading spaces (or whitespace in general, so tabs etc as well) on the
request line are silently accepted by some servers, like IIS. This
despite it being a RFC voilation.

The parsing code would consider the method fully parsed on the first
space it encountered, even if that was the first character of the line.
This leads to the method always being empty, the actual method then
being parsed as the uri, the uri as the http version. In case of IIS
this can be used to evade detection.

This patch first skips past leading whitespaces, and ignores it when
interpreting the method. It does issue a warning log message, so
libhtp users can be notified of the anomaly.