multi content: support multiple layers of encoding
Support multiple layers of compression in response bodies. 2 layers seem
quite common in some proxy environments.

To do so, make htp_tx_t::htp_decompressor_t a list instead of a single
instance. This allows for stacked encoding.

Setting up of multiple content encoding is a slow path.

Add a limit to how many layers of compressed data will be uncompressed.
Defaults to 2, but is configurable. Setting it to 0 disables the limit.

If the limit is reached, a warning will be recorded.