chunked: handle header/body mismatch
HTTP/1.0 and below don't support chunked encoding, however (some)
browsers behave differently. So try to decode the chunks in that case.

Other browsers however, will ignore the chunked header and treat the
body as unchunked.

This patch continues to try to dechunk the body in this case, but also
tries to detect wether the body is actually chunked. If not, for
HTTP/1.0 it will fall back to non-chunked processing of the body.