compression: rfc1950 delate & very liberal parsing
Implement rfc1950 deflate support.

If gzip or deflate is set, but data isn't compressed, pass it through
to the body callback unprocessed.

Changes the logic of the zlib_initialized field to mean which 'format'
was used.

Also if decompression fails, retry other methods. Browsers generally
accept it when the claimed method (e.g. deflate) is not actually
served (e.g. gzip).