finalize: handles requests and responses across packets
The problem fixed here comes with multiple requests in the
same session, when the second (or later) request is split over
multiple chunks, and more precisely when the HTTP method in the
request line gets splitted over multiple chunks.

In this case, htp_connp_REQ_FINALIZE probed to recongize a known
HTTP method, and failed, as for instance it only analyzed "GE"
when the final T for "GET" is in the next chunk.

The logic is now to peek enough data to have a full request line.
This way, we can decide if this input is the beginning of a new
request, or if it is the junk body of the previous one.

Pay attention to finalize dangling request if we see a response
after the current response.